Add required provider files for Flux.
We are adding these manually because they were not included in the 1.0.0
release of the Flux provider and are required in order for the docs
build to pass and have the provider added to the registry. Further
manual additions should not be necessary as the files in this PR have
been merged to the main branch of the Flux provider repo.