Do not clone the registry on build
As currently set up, running `resourcedocsgen docs registry` clones the registry (even
though `resourcedocsgen` is always called from the registry), then optionally checks out a
specific commit (how we run in CI), then builds docs there, then copies them over.

We don't have to do any of this. By assuming that `resourcedocsgen` is always run in the
registry root at the version, targeting the version of the registry we are checked out
as (which is always true in CI), we can save ourselves a full clone of the registry and
make developing the registry that much simpler.

This change also makes possible to build and develop the registry in a fork or without
internet, since building no longer requires a clone and checkout of a commit from
`github.com/pulumi/registry`.