Run markdown validation in parallel with the registry build (#5898)
We run a markdown lint pass on each PR build, which takes about 2m20s to run (on the payed
CI runner). This comprises about 7% of the total CI time needed to validate a contributing
PR. We can and should run this in parallel with the "Build and deploy preview" job.