Bring documentation generation into the registry
Presently, the registry's primary function is hosting documentation for
Pulumi packages. A Pulumi package is offered by a *provider plugin*. The
`pulumi-resource-aws` provider plugin, for instance, offers the `aws`
package, and implements CRUD methods for managing resources in an AWS
cloud. Alongside CRUD methods, provider plugins also offer a *schema*
(exposed through a `GetSchema` method) that details the resources,
functions, types and so on available to consumers of that provider and
its offered package. These schema power numerous automated generation
processes such as producing Pulumi language SDKs (e.g. the `@pulumi/aws`
NPM package for writing AWS programs in NodeJS, or the `pulumi_aws` PyPI
package for doing the same in Python); or writing Pulumi programs (e.g.
when using `pulumi convert`). They also power the generation of the vast
majority of the documentation we host in the registry.

At a very high level, documentation generation ("docsgen") takes as
input some schema (which contains documentation entries for the items in
the schema -- resources, functions, types, etc.) and produces HTML
output. While the `tools/resourcedocsgen` command in this repository
encapsulates this workflow, much of the code powering its functionality
is imported from the `pulumi/pulumi` repository's `pkg/codegen/docs`
package.

Historically, `pulumi/pulumi`'s `pkg/codegen` package hosted code
generators for SDKs ("SDKgen") and programs ("Programgen") in NodeJS,
Python, Golang and so on. When documentation generation ("docsgen") was
added, it too was added to `pkg/codegen`, as `pkg/codegen/docs`, and all
made sense. When support for Java and YAML was added to Pulumi, these
languages were managed in their own `pulumi-java` and `pulumi-yaml`
repositories respectively. Moreover, at some point, .Net support was
broken out of `pulumi/pulumi` into the `pulumi-dotnet` repository. These
repositories, all needing core functionality from `pulumi/pulumi`, took
necessary (and arguably sensible) dependencies on it.

While per-language code generation is for the most part (!) logical,
predictable and isolated (quirks in generating NodeJS code, for
instance, do not typically have any bearing on generating Python code),
documentation generation is not. A prime example (no pun intended) is
examples. Documentation contains examples in each language, and these
must be rendered correctly. Names of resources and their associated
types, methods, etc. must be cased and rendered correctly depending on
the language chosen by the documentation consumer. Hand-written
documentation is unstructured, and wrangling it to produce consistently
good looking (or even working) HTML can be very tricky. To try and limit
docsgen's hacks to "just the unavoidable language independent" messes,
the various language packages were extended to encapsulate
language-specific concerns, such as "how should I name this resource's
type in language X". This was good, but required docsgen to take a
dependency on each supported language's implementation. For languages
living outside of `pulumi/pulumi`, such as Java and YAML, this
introduced a *circular dependency*:

* `pulumi-java` and `pulumi-yaml` depend on `pulumi/pulumi` in order to
  access core functionality for building language hosts and code
  generators.
* `pulumi/pulumi` (specifically `pkg/codegen/docs`) depends on
  `pulumi-java` and `pulumi-yaml` in order to access language-specific
  functionaity for correctly generating e.g. identifiers and examples in
  documentation.

These circular dependencies are a huge barrier to refactoring and
evolving the coupled codebases. Making a change to code in
`pulumi/pulumi` that is consumed by e.g. `pulumi-yaml` is not directly
possible -- in the best case, a brand new piece of code must instead be
introduced and migrated to before the old code can be removed. Even in
the absence of these very real problems, circular dependencies are a
smell that a better factoring of concerns is possible.

Thankfully, such a factoring is not only possible in this case, it is
fairly simple! The only consumer of docsgen *is* the registry. We can
just move the docsgen code into the `resourcedocsgen` tool directly.
Doing so will enable us to delete the copied code from `pulumi/pulumi`
and break the related circular dependencies. Moreover, as the only
consumer of this code, the registry (or rather, those who work on it) is
the primary driver of change to docsgen. Moving it here makes making
such changes much easier without having to handle the entirety of
`pulumi/pulumi` and its development processes.

Finally therefore we arrive at the meat of this change: moving docsgen
into `tools/resourcedocsgen`. In doing so, we take the opportunity to
make the following clean-ups and improvements:

* The `Context` type, which encapsulated the legacy and
  formerly-non-paralellisable `docsGenContext` type, has just been
  collapsed into/replaced by `docsGenContext`. Moreover, there is no
  need to worry about separately initialising a created context any more
  -- `NewContext` always produces a valid starting point. The relevant
  tests have been updated to reflect this.

* The `TestSDKGen` test helpers that were being used in `pulumi/pulumi`
  have been vendored (and in the process) into this repository.
  `TestSDKGen` is slated for removal in `pulumi/pulumi` (to be replaced
  by language conformance tests), so it serves no value to take a
  dependency on it here. Moreover, by vendoring and specialising to just
  the docsgen use case, the test code (isolated in
  `gen_package_test.go`) is a lot simpler and easier to follow.

* General formatting and documentation clean ups, such as the removal of
  file-level `//nolint:lll` pragmas which had led to sloppiness in some
  places.

* The `README.md` has been updated to document testing docsgen code, as
  well as adding and updating test cases/expected outputs.

For ease of review, this change is split up into two commits: the
primary commit that brings over all the code, and a second one that
copies over the `testdata` and test cases. The second represents a huge
number of lines but is largely mechanical, and has been somewhat
verified to be correct by running a `diff` against the copy that remains
in `pulumi/pulumi` (which will be deleted when this lands).