Handle legacy (name,repoName) pairs (#6372)
https://github.com/pulumi/registry/pull/6246 introduced validation that the name a
provider specifies matches the name found in the schema. For community providers, the
specified provider name is matched to the repo name without the `pulumi-` prefix. All but
one community provider passes this stricter level of scrutiny.

The Runpod provider (maintained by Runpod) does not comply with this scheme. The provider
is called "runpod" in the schema, but the repo is called "runpod/pulumi-runpod-native". To
avoid breaking an already accepted provider, this adds a special exception for Runpod.

Fixes #6371