Display resource lookup for Pulumi YAML
Each resource API page has a "Lookup existing <Name> Resource" section, which documents
how to use the `Foo.get` method in Typescript and Python, the `.Get` method in C#, Java
and the `GetResource` function in Go. It claims that this is impossible in Pulumi YAML,
even though it isn't post pulumi/pulumi-yaml#155. This change adds the correct content for
Pulumi YAML.

Fixes #4279