CI for `scripts`: Yarn, ESLint and Prettier (#6535)
* Delete `package-lock.json` in favor of `yarn.lock`

This repository uses `yarn` in CI, and has a `yarn.lock`. Given that, I'm defaulting to
using `yarn` over `npm`.

* Add `prettier` as a formatter for `scripts`

Prettier was chosen because it is popular and opinionated. A tab width of 4 was chosen
only because that seems to be what we are already using.

Formatting is enforced in CI.

* Add eslint as a linter for `scripts`

eslint was chosen because it is industry standard, and used in other TS and JS projects at
Pulumi.

* Apply `yarn run format`

This commit is entirely automated. No manual changes were added.

* Fix linter failures