Fix FunctionArgsOutputVersion for Java (#6640)
Java uses the same args type for both the regular and output versions of the function. This corrects docsgen to set the correct `FunctionArgsOutputVersion`. Additionally it fixes java emitting broken output versions when no output version should be generated.

Note: The actual docs shouldn't change because we've worked around this issue previously by using the regular `FunctionArgs` even for the output version in Java.