Add file targets for `bin/resourcedocsgen` and docs content (#6625)
This change simplifies the process for building docs for a provider to a single command:

```sh
make content/registry/packages/<package_name>
```

Because we are using proper file targets with make, this means we get efficient rebuilds.

This introduces a pinned dependency on `iwahbe/helpmakego`. No developer action is needed
to consume the new makefiles.