Ignore ParameterNotFound error when cleaning up SSM parameters (#6708)
The bucket cleanup started failing after https://github.com/pulumi/registry/pull/6656 landed because AWS SSM returns an error when trying to delete a non-existent parameter (sad). It seems that some of the parameters got manually cleaned up.
This change ignores the 404s and treats them as a successful operations.