Update bucketv2 migration guide for dependent resources (#6744)
This updates the migration guide based on customer feedback. A very
common scenario will be that users will have other non-bucket resources
that reference an output from the `Bucket` resource. In those cases if
you try to delete the `Bucket` resource from state the operation will
fail saying that the bucket `can't be safely delete because the
following resources depend on it`.

This updates the guide to include this scenario of having a separate
resource that depends on the Bucket.

re https://github.com/pulumi/pulumi-aws/issues/5131