Add legacy publisher handling for resource metadata generation (#6753)
This change removed the code for defaulting the publisher name if it's not specified: https://github.com/pulumi/registry/pull/6246.

We do want to require a publisher, but in order to not break existing providers this change adds an allow list for using the old mechanism of inferring the publisher name based on the github org.

Test by running the metadata generation on all third party packages:
```
~ go build -C ./tools/resourcedocsgen . && cat community-packages/package-list.json | jq -c '.include[]' | while read -r package; do
    repoSlug=$(echo "$package" | jq -r '.repoSlug')
    schemaFile=$(echo "$package" | jq -r '.schemaFile')
    repoName=$(echo "$repoSlug" | cut -d'/' -f2 | sed 's/^pulumi-//')

    # Special case for runpod-native
    if [ "$repoName" = "runpod-native" ]; then
        repoName="runpod"
    fi

    version=$(cat "themes/default/data/registry/packages/$repoName.yaml" | yq -r '.version')

    GITHUB_TOKEN=$(gh auth token) ./tools/resourcedocsgen/resourcedocsgen metadata from-github \
            --repoSlug "$repoSlug" \
            --version "$version" \
            --schemaFile "$schemaFile" 2>&1 | grep -q "Publisher is a required field on the schema" && echo "$repoSlug"
done
~
```