Handle "Cloud provider detected" error in cypress tests
Ignores "Cloud provider detected" error being thrown by vector.co/pixel.js script.
These analytics scripts try to detect and block traffic from cloud hosted
environments, which a GH runner technically is, in order to prevent abuse and filter
out "non-human" traffic. So this handles that error so we do not fail the tests on
that specific error.
This replicates a similar fix applied to the docs page:
https://github.com/pulumi/docs/pull/14702