Improve bucket cleanup with lifecycle rules (#7639)
The S3 bucket cleanup has been consistently flaky because it has to delete all files in a bucket before being able to delete a bucket (S3 doesn't let you delete non-empty buckets). That's a looot of API requests we're firing off towards S3, ultimately making the cleanup job regularly take 2+ hours.

Instead of cleaning out the bucket ourselves, we can use lifecycle policies to make AWS to the heavy lifting in the background. We also don't have to pay for the API requests that way (💵).

The overall flow is to upsert the lifecycle policy when a bucket needs to be cleaned up and add a tag indicating when this was done. Once 48 hours have passed, the bucket should've been cleaned out by AWS (lifecycle rules run once a day afaik) and we can try deleting it if it's empty.