Use Registry Service for schema lookup (#7663)
We're exhausting the 1000 requests per hour rate limit for GITHUB_TOKEN in GitHub actions: https://docs.github.com/en/rest/using-the-rest-api/rate-limits-for-the-rest-api?apiVersion=2022-11-28#primary-rate-limit-for-github_token-in-github-actions

We have 289 providers in the registry - 102 of these are backed by cloudfront - so there's still ~187 providers for which we're fetching the schema on every PR and main branch push. This means we can only have ~5 of these per hour.

This change makes use of the registry service as a first level cache. If the package exists in the registry - we download the schema from there instead and only fall back to the VCS otherewise. This should give us a drastically longer runway for GitHub rate limits.