Render resource description (if present).
In some cases, resources described in Provider's `schema.json` contain a
`description` field. This change propagates that value from the parsed
schema all the way to the template. Hence, when the template gets
rendered (and if a description field exists for that resource), it will
display the resource associated description.