Publish schema.json files
Right now we use schema.json to generate provider documentation, but we
don't actually publish the schema.json at an end point users can access.

This is unfortunate. The schema.json is very valuable for tooling. This
change publishes it and also adds a link to it in the resulting docs.